ui: Move camera picker to the settings popup
