+++ ui/src/components/settings.tsx
@@ -1,19 +1,16 @@
-import { useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  Drawer,
-  DrawerContent,
-  DrawerHeader,
-  DrawerTitle,
-} from "@/components/ui/drawer";
+import { Button } from "@/components/ui/button";
 import {
   Dialog,
   DialogContent,
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
+import {
+  Drawer,
+  DrawerContent,
+  DrawerHeader,
+  DrawerTitle,
+} from "@/components/ui/drawer";
 import {
   Form,
   FormControl,
@@ -22,14 +19,24 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
+import { Label } from "@/components/ui/label";
 import { useMediaQuery } from "@/hooks/use-media-query";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useCallback, useEffect, useState } from "react";
+import { useForm } from "react-hook-form";
+import { z } from "zod";
+import { Select } from "./ui/select";
 
 export interface StreamConfig {
   streamUrl: string;
   prompt?: any;
+  selectedDeviceId?: string;
+}
+
+interface VideoDevice {
+  deviceId: string;
+  label: string;
 }
 
 const DEFAULT_CONFIG: StreamConfig = {
@@ -40,12 +47,14 @@ const DEFAULT_CONFIG: StreamConfig = {
 interface StreamSettingsProps {
   open: boolean;
   onOpenChange: (open: boolean) => void;
+  onDeviceChange: (deviceId: string) => void;
   onSave: (config: StreamConfig) => void;
 }
 
 export function StreamSettings({
   open,
   onOpenChange,
+  onDeviceChange,
   onSave,
 }: StreamSettingsProps) {
   const isDesktop = useMediaQuery("(min-width: 768px)");
@@ -67,7 +76,7 @@ export function StreamSettings({
               <div className="mt-4">Stream Settings</div>
             </DialogTitle>
           </DialogHeader>
-          <ConfigForm config={config} onSubmit={handleSubmit} />
+          <ConfigForm config={config} onDeviceChange={onDeviceChange} onSubmit={handleSubmit} />
         </DialogContent>
       </Dialog>
     );
@@ -80,7 +89,7 @@ export function StreamSettings({
           <DrawerTitle>Stream Settings</DrawerTitle>
         </DrawerHeader>
         <div className="px-4">
-          <ConfigForm config={config} onSubmit={handleSubmit} />
+          <ConfigForm config={config} onDeviceChange={onDeviceChange} onSubmit={handleSubmit} />
         </div>
       </DrawerContent>
     </Drawer>
@@ -93,16 +102,53 @@ const formSchema = z.object({
 
 interface ConfigFormProps {
   config: StreamConfig;
+  onDeviceChange: (deviceId: string) => void;
   onSubmit: (config: StreamConfig) => void;
 }
 
-function ConfigForm({ config, onSubmit }: ConfigFormProps) {
+function ConfigForm({ config, onDeviceChange, onSubmit }: ConfigFormProps) {
   const [prompt, setPrompt] = useState<any>(null);
+  const [videoDevices, setVideoDevices] = useState<VideoDevice[]>([]);
+  const [selectedDevice, setSelectedDevice] = useState<string>("");
+
+  useEffect(() => {
+    onDeviceChange(selectedDevice);
+  }, [selectedDevice]);
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: config,
   });
+
+  const getVideoDevices = useCallback(async () => {
+    try {
+      await navigator.mediaDevices.getUserMedia({ video: true });
+
+      const devices = await navigator.mediaDevices.enumerateDevices();
+      const videoDevices = devices
+        .filter(device => device.kind === 'videoinput')
+        .map(device => ({
+          deviceId: device.deviceId,
+          label: device.label || `Camera ${device.deviceId.slice(0, 5)}...`
+        }));
+
+      setVideoDevices(videoDevices);
+      if (videoDevices.length > 0) {
+        setSelectedDevice(curr => curr || videoDevices[0].deviceId);
+      }
+    } catch (err) {
+      console.error('Failed to get video devices');
+    }
+  }, []);
+
+  useEffect(() => {
+    getVideoDevices();
+    navigator.mediaDevices.addEventListener('devicechange', getVideoDevices);
+
+    return () => {
+      navigator.mediaDevices.removeEventListener('devicechange', getVideoDevices);
+    };
+  }, [getVideoDevices]);
 
   const handleSubmit = (values: z.infer<typeof formSchema>) => {
     onSubmit({
@@ -111,6 +157,7 @@ function ConfigForm({ config, onSubmit }: ConfigFormProps) {
         ? values.streamUrl.replace(/\/+$/, "")
         : values.streamUrl,
       prompt,
+      selectedDeviceId: selectedDevice,
     });
   };
 
@@ -151,6 +198,25 @@ function ConfigForm({ config, onSubmit }: ConfigFormProps) {
             accept=".json"
             onChange={handlePromptChange}
           ></Input>
+        </div>
+
+        <div className="mt-6 mb-4">
+          <Label>Camera</Label>
+          <Select
+            value={selectedDevice}
+            onValueChange={setSelectedDevice}
+          >
+            <Select.Trigger className="w-full mt-2">
+              {videoDevices.find(d => d.deviceId === selectedDevice)?.label || 'Select camera'}
+            </Select.Trigger>
+            <Select.Content>
+              {videoDevices.map((device) => (
+                <Select.Option key={device.deviceId} value={device.deviceId}>
+                  {device.label}
+                </Select.Option>
+              ))}
+            </Select.Content>
+          </Select>
         </div>
 
         <Button type="submit" className="w-full mt-4 mb-4">
